Started serialisation of map geom file
