+++ bspsuite-core/src/io/mod.rs
@@ -1,1 +1,4 @@
+pub mod map_geometry_file;
 pub mod map_source_file;
+
+mod json_utils;
